Update code for clarity
